adding tests to bq and separating gemini logic
