+++ ai_bubbles/package.json
@@ -3,7 +3,8 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest",
+    "start": "node --inspect node_modules/.bin/functions-framework --target=ytarfGET --signature-type=http"
   },
   "author": "",
   "license": "ISC",
@@ -16,5 +17,14 @@
     "@google/generative-ai": "^0.21.0",
     "json-2-csv": "^5.5.7",
     "tmp": "^0.2.3"
+  },
+  "devDependencies": {
+    "eslint": "^8.57.0",
+    "jest": "^29.4.3"
+  },
+  "jest": {
+    "clearMocks": true,
+    "resetMocks": true,
+    "restoreMocks": true
   }
 }
